Changement description des pages 2 et 3, pour test

+++ deploy-beta.sh
@@ -2,17 +2,24 @@ echo "Entrez le descriptif des changements et tapez [ENTER]: "
 read commitmsg
 set -e
 GITURL=https://github.com/gooodhub/goood-site-dev.git
-echo "npm install"
-npm install
-NODE_ENV=production node index.js
+if [ -d "./dist" ]
+then
 cd dist
 rm -rf .git/
+else
+mkdir dist
+cd dist
+fi
 git init
 git remote add origin $GITURL
 echo "git pull"
 git pull origin master:gh-pages
 echo "rm "
 ls -a | grep -v '.*' | xargs rm -rf
+cd ..
+echo "npm install"
+npm install
+NODE_ENV=production node index.js
 echo "git add"
 git add .
 git commit -am "$commitmsg"
